𝚆𝙴𝙻𝙲𝙾𝙼𝙴 𝚃𝚄𝚁𝙽 𝙾𝙵𝙵/𝙾𝙽 𝙵𝙸𝚇𝙴𝙳 🦠

+++ LuciferMoringstar_Robot/modules/welcome.py
@@ -31,17 +31,18 @@ from pyrogram.errors import ChatWriteForbidden
 
 @lucifermoringstar_robot.on_message(filters.group & filters.new_chat_members)
 async def welcome(client, update):
-    try:
+    settings = await get_settings(update.chat.id)
+    if settings["welcome"]:
         try:
-            settings = await get_settings(update.chat.id)
-            welcometext = settings["welcometext"]
-            new_members = update.from_user.mention
-            dell = await update.reply_text(welcometext.format(first_name = update.from_user.first_name, last_name = update.from_user.last_name, username = f"@{update.from_user.username}" or None, group_name = update.chat.title, mention = new_members), disable_web_page_preview=True)
-            await asyncio.sleep(1000)
-            await dell.delete()
-        except ChatWriteForbidden:
-            pass
-        except Exception as error:
-            pass
-    except Exception as error:       
-        await update.reply_text(f"{error}")
+            try:            
+                welcometext = settings["welcometext"]
+                new_members = update.from_user.mention
+                dell = await update.reply_text(welcometext.format(first_name = update.from_user.first_name, last_name = update.from_user.last_name, username = f"@{update.from_user.username}" or None, group_name = update.chat.title, mention = new_members), disable_web_page_preview=True)
+                await asyncio.sleep(1000)
+                await dell.delete()
+            except ChatWriteForbidden:
+                pass
+            except Exception as error:
+                pass
+        except Exception as error:       
+            await update.reply_text(f"{error}")
